jobs: deprecate functions needed for 20.1 only
As we improve the jobs code it turns out certain functions are used
only in test code or in the deprecated old lease adoption.

Release note: none.
