+++ pkg/jobs/deprecated.go
@@ -12,6 +12,7 @@ package jobs
 
 import (
 	"context"
+	"fmt"
 	"math/rand"
 	"time"
 
@@ -26,6 +27,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/errors"
+	"github.com/opentracing/opentracing-go"
 )
 
 // Functions deprecated in release 20.2 that should be removed in release 21.1.
@@ -256,7 +258,7 @@ WHERE status IN ($1, $2, $3, $4, $5) ORDER BY created DESC`
 			continue
 		}
 		// Adopt job and resume/revert it.
-		if err := job.adopt(ctx, payload.Lease); err != nil {
+		if err := job.deprecatedAdopt(ctx, payload.Lease); err != nil {
 			r.unregister(*id)
 			return errors.Wrap(err, "unable to acquire lease")
 		}
@@ -268,7 +270,7 @@ WHERE status IN ($1, $2, $3, $4, $5) ORDER BY created DESC`
 			return err
 		}
 		log.Infof(ctx, "job %d: resuming execution", *id)
-		errCh, err := r.resume(resumeCtx, resumer, resultsCh, job, nil)
+		errCh, err := r.deprecatedResume(resumeCtx, resumer, resultsCh, job, nil)
 		if err != nil {
 			r.unregister(*id)
 			return err
@@ -331,4 +333,124 @@ func (r *Registry) deprecatedCancelAll(ctx context.Context) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	r.deprecatedCancelAllLocked(ctx)
+}
+
+// deprecatedResume starts or resumes a job. If no error is returned then the
+// job was asynchronously executed. The job is executed with the ctx, so ctx
+// must only by canceled if the job should also be canceled. resultsCh is passed
+// to the resumable func and should be closed by the caller after errCh sends
+// a value. The onDone function is called when the async task completes or if
+// an error is returned.
+func (r *Registry) deprecatedResume(
+	ctx context.Context, resumer Resumer, resultsCh chan<- tree.Datums, job *Job, onDone func(),
+) (<-chan error, error) {
+	errCh := make(chan error, 1)
+	if err := r.stopper.RunAsyncTask(ctx, job.taskName(), func(ctx context.Context) {
+		if onDone != nil {
+			defer onDone()
+		}
+		// Bookkeeping.
+		payload := job.Payload()
+		phs, cleanup := r.planFn("resume-"+job.taskName(), payload.Username)
+		defer cleanup()
+		spanName := fmt.Sprintf(`%s-%d`, payload.Type(), *job.ID())
+		var span opentracing.Span
+		ctx, span = r.ac.AnnotateCtxWithSpan(ctx, spanName)
+		defer span.Finish()
+
+		// Run the actual job.
+		status, err := job.CurrentStatus(ctx)
+		if err == nil {
+			var finalResumeError error
+			if job.Payload().FinalResumeError != nil {
+				finalResumeError = errors.DecodeError(ctx, *job.Payload().FinalResumeError)
+			}
+			err = r.stepThroughStateMachine(ctx, phs, resumer, resultsCh, job, status, finalResumeError)
+			if err != nil {
+				// TODO (lucy): This needs to distinguish between assertion errors in
+				// the job registry and assertion errors in job execution returned from
+				// Resume() or OnFailOrCancel(), and only fail on the former. We have
+				// tests that purposely introduce bad state in order to produce
+				// assertion errors, which shouldn't cause the test to panic. For now,
+				// comment this out.
+				// if errors.HasAssertionFailure(err) {
+				// 	log.ReportOrPanic(ctx, nil, err.Error())
+				// }
+				log.Errorf(ctx, "job %d: adoption completed with error %v", *job.ID(), err)
+			}
+			status, err := job.CurrentStatus(ctx)
+			if err != nil {
+				log.Errorf(ctx, "job %d: failed querying status: %v", *job.ID(), err)
+			} else {
+				log.Infof(ctx, "job %d: status %s after adoption finished", *job.ID(), status)
+			}
+		}
+		r.unregister(*job.ID())
+		errCh <- err
+	}); err != nil {
+		if onDone != nil {
+			onDone()
+		}
+		return nil, err
+	}
+	return errCh, nil
+}
+
+func (j *Job) deprecatedInsert(ctx context.Context, id int64, lease *jobspb.Lease) error {
+	if j.id != nil {
+		// Already created - do nothing.
+		return nil
+	}
+
+	j.mu.payload.Lease = lease
+
+	if err := j.runInTxn(ctx, func(ctx context.Context, txn *kv.Txn) error {
+		// Note: although the following uses ReadTimestamp and
+		// ReadTimestamp can diverge from the value of now() throughout a
+		// transaction, this may be OK -- we merely required ModifiedMicro
+		// to be equal *or greater* than previously inserted timestamps
+		// computed by now(). For now ReadTimestamp can only move forward
+		// and the assertion ReadTimestamp >= now() holds at all times.
+		j.mu.progress.ModifiedMicros = timeutil.ToUnixMicros(txn.ReadTimestamp().GoTime())
+		payloadBytes, err := protoutil.Marshal(&j.mu.payload)
+		if err != nil {
+			return err
+		}
+		progressBytes, err := protoutil.Marshal(&j.mu.progress)
+		if err != nil {
+			return err
+		}
+
+		if j.createdBy == nil {
+			const stmt = "INSERT INTO system.jobs (id, status, payload, progress) VALUES ($1, $2, $3, $4)"
+			_, err = j.registry.ex.Exec(ctx, "job-insert", txn, stmt, id, StatusRunning, payloadBytes, progressBytes)
+			return err
+		}
+		const stmt = `
+INSERT INTO system.jobs (id, status, payload, progress, created_by_type, created_by_id) 
+VALUES ($1, $2, $3, $4, $5, $6)`
+		_, err = j.registry.ex.Exec(ctx, "job-insert", txn, stmt,
+			id, StatusRunning, payloadBytes, progressBytes, j.createdBy.Name, j.createdBy.ID)
+		return err
+
+	}); err != nil {
+		return err
+	}
+	j.id = &id
+	return nil
+}
+
+func (j *Job) deprecatedAdopt(ctx context.Context, oldLease *jobspb.Lease) error {
+	return j.Update(ctx, func(txn *kv.Txn, md JobMetadata, ju *JobUpdater) error {
+		if !md.Payload.Lease.Equal(oldLease) {
+			return errors.Errorf("current lease %v did not match expected lease %v",
+				md.Payload.Lease, oldLease)
+		}
+		md.Payload.Lease = j.registry.deprecatedNewLease()
+		if md.Payload.StartedMicros == 0 {
+			md.Payload.StartedMicros = timeutil.ToUnixMicros(j.registry.clock.Now().GoTime())
+		}
+		ju.UpdatePayload(md.Payload)
+		return nil
+	})
 }
