+++ pkg/jobs/helpers_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/errors"
 )
 
 // FakeResumer calls optional callbacks during the job lifecycle.
@@ -69,10 +70,12 @@ func (j *Job) Started(ctx context.Context) error {
 	return j.started(ctx)
 }
 
-// Created is a wrapper around the internal function that creates the initial
-// job state.
+// Created is a test only function that inserts a new jobs table row.
 func (j *Job) Created(ctx context.Context) error {
-	return j.created(ctx)
+	if j.ID() != nil {
+		return errors.Errorf("job already created with ID %v", *j.ID())
+	}
+	return j.deprecatedInsert(ctx, j.registry.makeJobID(), nil /* lease */)
 }
 
 // Paused is a wrapper around the internal function that moves a job to the
